requirements.txt
absl-py==0.10.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
bleach==3.2.1
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
entrypoints==0.3
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
importlib-resources==3.3.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
pyzmq==20.0.0
qtconsole==5.0.0
QtPy==1.9.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.4.1
seaborn==0.11.0
Send2Trash==1.5.0
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow-addons==0.11.2
tensorflow-datasets==4.1.0
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0
tensorflow-metadata==0.25.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tifffile==2020.9.3
toml==0.10.2
tornado==6.1
tqdm==4.51.0
traitlets==4.3.3
typeguard==2.10.0
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
